% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{new.experiment}
\alias{new.experiment}
\title{New R-API MaveDB Experiment}
\usage{
new.experiment(data)
}
\value{
a new R-API MaveDB experiment object.
}
\description{
Internal constructor that creates an immutable new Experiment object. This function is not exported
  and should not be called outside of the rapimave package itself. Experiment objects are generated
  by the \code{getAllExperiments()} and \code{getExperiment()} methods of a rapimave object (which itself
 is generated using the \code{new.rapimave()} constructor). However, this section explains the
  functions that are available for these objects.
}
\details{
A Experiment object offers the following getter functions:
\itemize{
  \item{getScoreSets()} return the list of accessions of the Experiment's ScoreSets.
  \item{getAuthors()} returns the usernames of the authors of this Experiment.
  \item{getAccession()} return the accession of this Experiment.
  \item{getExperimentSet()} return the accession of the ExperimentSet to which this Experiment belongs.
}
}
\examples{
\dontrun{
mave <- new.rapimave()
exp1 <- mave$getExperiment("EXP000001A")
authorIds <- exp1$getAuthors()
}
}
