% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{new.experimentSet}
\alias{new.experimentSet}
\title{New R-API MaveDB ExperimentSet}
\usage{
new.experimentSet(data)
}
\value{
a new R-API MaveDB ExperimentSet object.
}
\description{
Internal constructor that creates an immutable new ExperimentSet object. This function is not exported
  and should not be called outside of the rapimave package itself. ExperimentSet objects are generated
  by the \code{getAllExperimentSets()} and \code{getExperimentSet()} methods of a rapimave object (which itself
 is generated using the \code{new.rapimave()} constructor). However, this section explains the
  functions that are available for these objects.
}
\details{
A ExperimentSet object offers the following getter functions:
\itemize{
  \item{getContributors()} returns a list of users that contributed to this record.
  \item{getURN()} returns the URN of this record.
  \item{getCreationDate()} returns the date when this record was originally created.
  \item{getModificationDate()} returns the date when this record was last modified.
  \item{getPublicationDate()} returns the date when this record was published.
  \item{getCreatedBy()} returns the user who originally created this record.
  \item{getModifiedBy()} returns the user who last modified this record.
  \item{getMetadata()} returns the metadata for this record.
  \item{getTitle()} returns the title of this record.
  \item{getShortDescription()} returns a short description text of this record.
  \item{getAbstract()} returns the abstract text of this record.
  \item{getMethods()} returns the methods description text of this record.
  \item{getKeywords()} returns the list of keywords associated with this record.
  \item{getDOIs()} returns the list of Digital Object Identifiers (DOIs) associated with this record.
  \item{getXrefSRA()} returns cross-references to the NCBI Short Read Archive (SRA).
  \item{getXRefPubmed()} returns cross-references to Pubmed.
  \item{getExperiments()} returns the URNs of the experiments belonging to this ExperimentSet.
}
}
\examples{
\dontrun{
mave <- new.rapimave()
set1 <- mave$getExperimentSet("urn:mavedb:00000001")
expURNs <- set1$getExperiments()
}
}
