% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{new.rapimave}
\alias{new.rapimave}
\title{MaveDB R-API client constructor}
\usage{
new.rapimave(baseURL = "https://www.mavedb.org/api/", certifySSL = FALSE,
  encoding = "UTF-8")
}
\arguments{
\item{baseURL}{MaveDB API base-URL. Defaults to "https://www.mavedb.org/api/"}

\item{certifySSL}{Logical indicating whether to certify SSL certificates when querying
the database. Defaults to \code{FALSE}.}

\item{encoding}{A string specifying the expected character encoding used by the webservice.
Defaults to "UTF-8"}
}
\value{
MaveDB R-API object with functions accessbile via the $ operator.
}
\description{
Creates an object of type rapimave.
}
\details{
The resulting object offers the following functions:
\itemize{
	\item{getAllUsers()} returns a list of all user objects. 
    See \code{\link{new.user}} for available methods on user objects.
	\item{getUser(username)} returns the user object for the given username. 
    Usernames are usually ORCID IDs. See https://orcid.org/ for information on ORCID.
    See \code{\link{new.user}} for available methods on user objects.
	\item{getAllExperimentSets()} returns a list of all ExperimentSet objects.
    See \code{\link{new.experimentSet}} for available methods on ExperimentSet objects.
	\item{getExperimentSet(accession)} returns the ExperimentSet object for the given accession.
    ExperimentSet accessions usually follow the syntax /EXPS\d+/ .
    See \code{\link{new.experimentSet}} for available methods on ExperimentSet objects.
	\item{getAllExperiments()} returns a list of all Experiment objects.
    See \code{\link{new.experiment}} for available methods on Experiment objects.
	\item{getExperiment(accession)} returns the Experiment object for the given accession.
    Experiment accessions usually follow the syntax /EXP\d+\w[1]/ .
    See \code{\link{new.experiment}} for available methods on Experiment objects.
	\item{getAllScoreSets()} returns a list of all ScoreSet objects.
    See \code{\link{new.scoreSet}} for available methods on ScoreSet objects.
	\item{getScoreSet(accession)} returns the ScoreSet object for the given accession.
    ScoreSet accessions usually follow the syntax /SCS\d+\w[1]\.\d[1]/ .
    See \code{\link{new.scoreSet}} for available methods on ScoreSet objects.
	\item{getScores(accession)} returns a \code{data.frame} with the scores for the given ScoreSet accession
	\item{getCounts(accession)} returns a \code{data.frame} with the counts for the given ScoreSet accession
}
}
\examples{
\dontrun{
mave <- new.rapimave()
}
}
\keyword{MaveDB}
