% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{new.refmap}
\alias{new.refmap}
\title{New R-API MaveDB ReferenceMap}
\usage{
new.refmap(data)
}
\value{
a new R-API MaveDB ReferenceMap object.
}
\description{
Internal constructor that creates an immutable new ReferenceMap object. This function is not exported
  and should not be called outside of the rapimave package itself. ReferenceMap objects are generated
  by the getReferenceMap() method of a target object (which itself
 is generated using the \code{getTarget()} method on a scoreset object). However, this section explains the
  functions that are available for these objects.
}
\details{
A ScoreSet object offers the following getter functions:
\itemize{
  \item{getGenome()} returns the genome object for this reference map
  \item{isPrimary()} returns whether this is the primary reference map for this scoreset
  \item{getIntervals()} returns a data.frame listing the intervals in the genome to which 
     the target is mapped.
}
}
