% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{new.target}
\alias{new.target}
\title{New R-API MaveDB Target}
\usage{
new.target(data)
}
\value{
a new R-API MaveDB Target object.
}
\description{
Internal constructor that creates an immutable new Target object. This function is not exported
  and should not be called outside of the rapimave package itself. Target objects are generated
  by the getTarget() method of a scoreset object (which itself
 is generated using the \code{getAllScoreSets()} and \code{getScoreSet()} methods of a rapimave 
  object). However, this section explains the
  functions that are available for these objects.
}
\details{
A ScoreSet object offers the following getter functions:
\itemize{
  \item{getName()} returns the name of the target
  \item{getSequence()} returns wildtype sequence of the target
  \item{getSequenceType()} returns whether the sequence is at DNA or Protein level
  \item{getXrefUniprot()} returns a Xref object for Uniprot
  \item{getXrefEnsembl()} returns a Xref object for Ensembl
  \item{getXrefRefseq()} returns a Xref object for Refseq
  \item{getReferenceMaps()} returns a list of reference map objects that describe how this 
     target maps to different reference genomes.
  \item{getScoreset()} retunrs the URN of the scoreset to which this target belongs.
}
}
