% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{new.user}
\alias{new.user}
\title{New R-API MaveDB User}
\usage{
new.user(data)
}
\value{
a new R-API MaveDB user object.
}
\description{
Internal constructor that creates an immutable new user object. This function is not exported
  and should not be called outside of the rapimave package itself. User objects are generated
  by the \code{getAllUsers()} and \code{getUser()} methods of a rapimave object (which itself
 is generated using the \code{new.rapimave()} constructor). However, this section explains the
  functions that are available for these objects.
}
\details{
A user object offers the following getter functions:
\itemize{
  \item{getUsername()} return the username.
  \item{getFirstName()} return the user's first name.
  \item{getLastName()} return the user's last name.
  \item{getDisplayName()} return the user's preferred display name.
  \item{getExperimentSets()} return the ExperimentSets attributed to the user.
  \item{getExperiments()} return the Experiments attributed to the user.
  \item{getScoreSets()} return the ScoreSets attributed to the user.
}
}
\examples{
\dontrun{
mave <- new.rapimave()
user <- mave$getUser("0000-0003-1474-605X")
cat(user$getFirstName(),user$getLastName())
}
}
