% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{new.xref}
\alias{new.xref}
\title{New R-API MaveDB cross-reference}
\usage{
new.xref(data)
}
\value{
a new R-API MaveDB Xref object.
}
\description{
Internal constructor that creates an immutable new Xref object. This function is not exported
  and should not be called outside of the rapimave package itself. Xref objects are generated
  by various methods of a rapimave object (which itself
 is generated using the \code{new.rapimave()} constructor). However, this section explains the
  functions that are available for these objects.
}
\details{
A ScoreSet object offers the following getter functions:
\itemize{
  \item{getID()} returns the Identifier or Accession of this cross-reference
  \item{getURL()} returns the database URL
  \item{getDB()} returns the database name
  \item{getDBVersion()} returns the database version
}
}
