% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{print.rapimave}
\alias{print.rapimave}
\title{Print R-API MaveDB object}
\usage{
\method{print}{rapimave}(obj)
}
\arguments{
\item{obj}{the object to print}
}
\description{
Prints a human-readable summary of a R-API MaveDB object.
}
\details{
Prints a list of available functions for the object.
}
\examples{
\dontrun{
mave <- new.rapimave()
print(mave)
}
}
\keyword{MaveDB}
