% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{print.rapimaveGenome}
\alias{print.rapimaveGenome}
\title{Print R-API MaveDB Genome object}
\usage{
\method{print}{rapimaveGenome}(obj)
}
\arguments{
\item{obj}{the object to print}
}
\description{
Prints a human-readable summary of a R-API MaveDB genome object:
}
\keyword{MaveDB}
\keyword{print}
