% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{print.rapimaveRefmap}
\alias{print.rapimaveRefmap}
\title{Print R-API MaveDB reference map object}
\usage{
\method{print}{rapimaveRefmap}(obj)
}
\arguments{
\item{obj}{the object to print}
}
\description{
Prints a human-readable summary of a R-API MaveDB reference map object:
}
\keyword{MaveDB}
\keyword{print}
