% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{print.rapimaveScoreSet}
\alias{print.rapimaveScoreSet}
\title{Print R-API MaveDB ScoreSet}
\usage{
\method{print}{rapimaveScoreSet}(obj)
}
\arguments{
\item{obj}{the ScoreSet object to print}
}
\description{
Prints a human-readable summary of a R-API MaveDB ScoreSet object:
}
\examples{
\dontrun{
mave <- new.rapimave()
print(mave$getScoreSet("SCS000001A.2"))
}
}
\keyword{MaveDB}
\keyword{ScoreSet}
