% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{print.rapimaveTarget}
\alias{print.rapimaveTarget}
\title{Print R-API MaveDB target object}
\usage{
\method{print}{rapimaveTarget}(obj)
}
\arguments{
\item{obj}{the object to print}
}
\description{
Prints a human-readable summary of a R-API MaveDB target object:
}
\keyword{MaveDB}
\keyword{print}
