% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{print.rapimaveUser}
\alias{print.rapimaveUser}
\title{Print R-API MaveDB User}
\usage{
\method{print}{rapimaveUser}(obj)
}
\arguments{
\item{obj}{the User object to print}
}
\description{
Prints a human-readable summary of a R-API MaveDB User object:
}
\examples{
\dontrun{
mave <- new.rapimave()
print(mave$getUser("0000-0003-1474-605X"))
}
}
\keyword{MaveDB}
