% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{print.rapimaveXref}
\alias{print.rapimaveXref}
\title{Print R-API MaveDB cross-reference}
\usage{
\method{print}{rapimaveXref}(obj)
}
\arguments{
\item{obj}{the object to print}
}
\description{
Prints a human-readable summary of a R-API MaveDB cross-reference object:
}
\keyword{MaveDB}
\keyword{print}
