% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{new.experimentSet}
\alias{new.experimentSet}
\title{New R-API MaveDB ExperimentSet}
\usage{
new.experimentSet(data)
}
\value{
a new R-API MaveDB ExperimentSet object.
}
\description{
Internal constructor that creates an immutable new ExperimentSet object. This function is not exported
  and should not be called outside of the rapimave package itself. ExperimentSet objects are generated
  by the \code{getAllExperimentSets()} and \code{getExperimentSet()} methods of a rapimave object (which itself
 is generated using the \code{new.rapimave()} constructor). However, this section explains the
  functions that are available for these objects.
}
\details{
A ExperimentSet object offers the following getter functions:
\itemize{
  \item{getExperiments()} returns the accessions of the experiments belonging to this ExperimentSet.
  \item{getAuthors()} returns the usernames of the authors of this ExperimentSet.
  \item{getAccession()} returns the accession of this ExperimentSet.
}
}
\examples{
\dontrun{
mave <- new.rapimave()
set1 <- mave$getExperimentSet("EXPS000001")
expAccessions <- set1$getExperiments()
}
}
