% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{new.rapimave}
\alias{new.rapimave}
\title{MaveDB R-API client constructor}
\usage{
new.rapimave(
  baseURL = "https://www.mavedb.org/api/",
  certifySSL = FALSE,
  encoding = "UTF-8"
)
}
\arguments{
\item{baseURL}{MaveDB API base-URL. Defaults to "https://www.mavedb.org/api/"}

\item{certifySSL}{Logical indicating whether to certify SSL certificates when querying
the database. Defaults to \code{FALSE}.}

\item{encoding}{A string specifying the expected character encoding used by the webservice.
Defaults to "UTF-8"}
}
\value{
MaveDB R-API object with functions accessbile via the $ operator.
}
\description{
Creates an object of type rapimave.
}
\details{
The resulting object offers the following functions:
\itemize{
	\item{getAllUsers()} returns a list of all user objects. 
    See \code{\link{new.user}} for available methods on user objects.
	\item{getUser(username)} returns the user object for the given username. 
    Usernames are usually ORCID IDs. See \url{https://orcid.org/} for information on ORCID.
    See \code{\link{new.user}} for available methods on user objects.
	\item{findExperimentSets(filter)} returns a list of all ExperimentSet objects that match the given filter word.
    See \code{\link{new.experimentSet}} for available methods on ExperimentSet objects.
	\item{getAllExperimentSets()} returns a list of all ExperimentSet objects.
    See \code{\link{new.experimentSet}} for available methods on ExperimentSet objects.
	\item{getExperimentSet(urn)} returns the ExperimentSet object for the given URN.
    ExperimentSet URNs usually follow the syntax /^urn:mavedb:\\d+$/ .
    See \code{\link{new.experimentSet}} for available methods on ExperimentSet objects.
	\item{getAllExperiments()} returns a list of all Experiment objects.
    See \code{\link{new.experiment}} for available methods on Experiment objects.
	\item{findExperiments(filter)} returns a list of all Experiment objects that match the given filter word.
    See \code{\link{new.experiment}} for available methods on Experiment objects.
	\item{getExperiment(urn)} returns the Experiment object for the given URN.
    Experiment URNs usually follow the syntax /^urn:mavedb:\\d+-\\w+/ .
    See \code{\link{new.experiment}} for available methods on Experiment objects.
	\item{getAllScoreSets()} returns a list of all ScoreSet objects.
    See \code{\link{new.scoreSet}} for available methods on ScoreSet objects.
	\item{findScoreSets(filter)} returns a list of all ScoreSet objects that match the given filter word.
    See \code{\link{new.scoreSet}} for available methods on ScoreSet objects.
	\item{getScoreSet(urn)} returns the ScoreSet object for the given URN.
    ScoreSet URNs usually follow the syntax /^urn:mavedb:\\d+-\\w+-\\d+$/ .
    See \code{\link{new.scoreSet}} for available methods on ScoreSet objects.
	\item{getScores(urn)} returns a \code{data.frame} with the scores for the given ScoreSet URN
	\item{getCounts(urn)} returns a \code{data.frame} with the counts for the given ScoreSet URN
	\item{getMetadata(urn)} returns a \code{data.frame} with the metadata for the given ScoreSet URN
 \item{getKeywords()} returns all currently used keywords across MaveDB.
 \item{getAllPubmed()} returns all currently used Pubmed IDs across MaveDB.
 \item{getAllDOI()} returns all currently used DOIs across MaveDB.
 \item{getAllSRA()} returns all currently used SRA IDs across MaveDB.
 \item{getAllRefseq()} returns all currently used Refseq IDs across MaveDB.
 \item{getAllUniprot()} returns all currently used Uniprot IDs across MaveDB.
 \item{getAllEnsembl()} returns all currently used Ensembl IDs across MaveDB.
 \item{getAllAssemblies()} returns all currently used Genome assemblies across MaveDB.
 \item{getAllTargets()} returns all currently used target sequences across MaveDB.
 \item{getReferenceGenomes()} returns all currently used reference genomes across MaveDB.
}
}
\examples{
\dontrun{
mave <- new.rapimave()
}
}
\keyword{MaveDB}
