% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapimave.R
\name{new.scoreSet}
\alias{new.scoreSet}
\title{New R-API MaveDB ScoreSet}
\usage{
new.scoreSet(data)
}
\value{
a new R-API MaveDB ScoreSet object.
}
\description{
Internal constructor that creates an immutable new ScoreSet object. This function is not exported
  and should not be called outside of the rapimave package itself. ScoreSet objects are generated
  by the \code{getAllScoreSets()} and \code{getScoreSet()} methods of a rapimave object (which itself
 is generated using the \code{new.rapimave()} constructor). However, this section explains the
  functions that are available for these objects.
}
\details{
A ScoreSet object offers the following getter functions:
\itemize{
  \item{getAuthors()} returns the usernames of the authors of this ScoreSet.
  \item{getAccession()} returns the accession of this ScoreSet.
  \item{getLicence()} returns the licencing information governing the use of this ScoreSet.
  \item{getCurrentVersion()} If a more up-to-date version of this ScoreSet exists, 
    that has since replaced it, this function will return the accession of the most recent version.
  \item{getReplaces()} If this ScoreSet replaces an older ScoreSet, this function returns
    the accession of that older ScoreSet.
  \item{getReplacedBy()} If this ScoreSet was replaced by newer ScoreSet, this function returns
    the accession of that newer ScoreSet in versioning history. To obtain the most recent version,
    use \code{getCurrentVersion()}.
  \item{getCountColumns()} returns the column names in the count table for this ScoreSet.
  \item{getScoreColumns()} returns the column names in the score table for this ScoreSet.
}
}
\examples{
\dontrun{
mave <- new.rapimave()
set1 <- mave$getScoreSet("SCS000001A.2")
scoreColNames <- set1$getScoreColumns()
}
}
